# docs/openapi.yaml
openapi: 3.0.3
info:
  title: YES LINKS API
  version: 0.1.0
  description: URL shortener with redirects, click tracking and analytics.
servers:
  - url: https://y.es
  - url: http://localhost:8001
paths:
  /health:
    get:
      summary: Health check
      operationId: health
      responses:
        '200':
          description: Service health
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /links:
    get:
      summary: List links
      operationId: listLinks
      parameters:
        - in: query
          name: campaign
          schema:
            type: string
          description: Filter by campaign (exact match)
        - in: query
          name: tags
          schema:
            type: string
          description: Comma-separated tag list — all must be present on each link
        - in: query
          name: search
          schema:
            type: string
          description: Substring search on short_code or target_url
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Paginated link list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinksListResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create short link
      operationId: createLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkRequest'
      responses:
        '201':
          description: Link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dashboard/summary:
    get:
      summary: Global KPI summary
      operationId: getDashboardSummary
      tags: [dashboard]
      responses:
        '200':
          description: KPI summary with trends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummaryResponse'
  /campaigns/stats:
    get:
      summary: Per-campaign aggregation
      operationId: getCampaignsStats
      tags: [dashboard]
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
          description: ISO 8601 start datetime (inclusive)
        - in: query
          name: to
          schema:
            type: string
            format: date-time
          description: ISO 8601 end datetime (exclusive)
      responses:
        '200':
          description: Campaign stats list ordered by total_clicks desc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignStatsRow'
  /{short_code}:
    get:
      summary: Redirect by short code
      operationId: redirect
      parameters:
        - in: path
          name: short_code
          required: true
          schema:
            type: string
            minLength: 5
            maxLength: 8
      responses:
        '302':
          description: Redirect to target URL
          headers:
            Location:
              schema:
                type: string
        '404':
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /links/{id}/stats:
    get:
      summary: Link analytics
      operationId: getLinkStats
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '404':
          description: Link not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /links/{id}:
    put:
      summary: Update link
      operationId: updateLink
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLinkRequest'
      responses:
        '200':
          description: Link updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '404':
          description: Link not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete link
      operationId: deleteLink
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteLinkResponse'
        '404':
          description: Link not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
    CreateLinkRequest:
      type: object
      required: [target_url]
      properties:
        target_url:
          type: string
          format: uri
        campaign:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
    UpdateLinkRequest:
      type: object
      required: [target_url]
      properties:
        target_url:
          type: string
          format: uri
        campaign:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
    LinkResponse:
      type: object
      required: [id, short_code, short_url, target_url, tags, created_at, created_by]
      properties:
        id:
          type: string
        short_code:
          type: string
          minLength: 5
          maxLength: 8
        short_url:
          type: string
          format: uri
        target_url:
          type: string
          format: uri
        campaign:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
    ClickByDay:
      type: object
      required: [day, count]
      properties:
        day:
          type: string
        count:
          type: integer
    ClickByCampaign:
      type: object
      required: [campaign, count]
      properties:
        campaign:
          type: string
        count:
          type: integer
    StatsResponse:
      type: object
      required: [link_id, total_clicks, clicks_by_day, clicks_by_campaign]
      properties:
        link_id:
          type: string
        total_clicks:
          type: integer
        clicks_by_day:
          type: array
          items:
            $ref: '#/components/schemas/ClickByDay'
        clicks_by_campaign:
          type: array
          items:
            $ref: '#/components/schemas/ClickByCampaign'
    DeleteLinkResponse:
      type: object
      required: [id, deleted]
      properties:
        id:
          type: string
        deleted:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
    LinksListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LinkResponse'
        total:
          type: integer
          description: Total count before pagination
    TopCampaign:
      type: object
      required: [campaign, clicks]
      properties:
        campaign:
          type: string
        clicks:
          type: integer
    ClickTrends:
      type: object
      required: [clicks_last_7d, clicks_prev_7d, pct_change]
      properties:
        clicks_last_7d:
          type: integer
        clicks_prev_7d:
          type: integer
        pct_change:
          type: number
          nullable: true
          description: Percentage change vs previous 7-day window (null if no prior data)
    DashboardSummaryResponse:
      type: object
      required: [total_links, total_clicks, avg_clicks_per_link, top_campaigns, trends]
      properties:
        total_links:
          type: integer
        total_clicks:
          type: integer
        avg_clicks_per_link:
          type: number
        top_campaigns:
          type: array
          items:
            $ref: '#/components/schemas/TopCampaign'
          description: Top 5 campaigns by click count
        trends:
          $ref: '#/components/schemas/ClickTrends'
    CampaignStatsRow:
      type: object
      required: [campaign, total_links, total_clicks]
      properties:
        campaign:
          type: string
        total_links:
          type: integer
          description: Number of links tagged with this campaign (including 0-click links)
        total_clicks:
          type: integer
        last_click_at:
          type: string
          format: date-time
          nullable: true